+++ Application/EdFi.Ods.AdminApp.Web/Controllers/OdsInstanceSettingsController.cs
@@ -153,94 +153,6 @@ namespace EdFi.Ods.AdminApp.Web.Controllers
             return View(model);
         }
 
-        public async Task<ActionResult> LearningStandards()
-        {
-            var odsApiFacade = await _odsApiFacadeFactory.Create();
-            var config = await _odsSecretConfigurationProvider.GetSecretConfiguration(_instanceContext.Id);
-
-            if (config == null)
-            {
-                _logger.Error("ODS secret configuration is null.");
-                throw new InvalidOperationException("ODS secret configuration can not be null.");
-            }
-
-            var model = new OdsInstanceSettingsModel
-            {
-                OdsInstanceSettingsTabEnumerations =
-                    _tabDisplayService.GetOdsInstanceSettingsTabDisplay(OdsInstanceSettingsTabEnumeration
-                        .LearningStandards),
-                LearningStandardsModel = new LearningStandardsModel
-                {
-                    ApiKey = config.LearningStandardsCredential?.ApiKey ?? string.Empty,
-                    ApiSecret = config.LearningStandardsCredential?.ApiSecret ?? string.Empty,
-                    HasApiData = odsApiFacade.DoesApiDataExist(),
-                    SynchronizationWasSuccessful =
-                        config.LearningStandardsCredential?.SynchronizationWasSuccessful ?? false,
-                    IsJobRunning = _learningStandardsJob.IsJobRunning(),
-                    IsSameOdsInstance = _learningStandardsJob.IsSameOdsInstance(_instanceContext.Id, typeof(LearningStandardsJobContext)),
-                    LastUpdatedDate = config.LearningStandardsCredential?.LastUpdatedDate ?? DateTime.MinValue
-                },
-                OdsInstance = _instanceContext
-            };
-
-            return View(model);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> LearningStandards(OdsInstanceSettingsModel model)
-        {
-            await _learningStandardsSetupCommand.Execute(
-                new AcademicBenchmarkConfig
-                {
-                    ApiKey = model.LearningStandardsModel.ApiKey,
-                    ApiSecret = model.LearningStandardsModel.ApiSecret,
-                    OdsApiMode = CloudOdsAdminAppSettings.Instance.Mode
-                });
-
-            await RunLearningStandardsJob();
-
-            return Ok();
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> UpdateLearningStandards()
-        {
-            await RunLearningStandardsJob();
-
-            return Ok();
-        }
-
-        private async Task RunLearningStandardsJob()
-        {
-            var connectionInformation = await GetConnectionInformationProvider();
-
-            int? schoolYear = null;
-
-            if (_cloudOdsAdminAppSettingsApiModeProvider.GetApiMode() == ApiMode.YearSpecific)
-            {
-                schoolYear = _instanceContext.Name.ExtractNumericInstanceSuffix();
-            }
-
-            var jobContext = new LearningStandardsJobContext
-            {
-                ApiUrl = connectionInformation.ApiServerUrl,
-                SchoolYear = schoolYear,
-                OdsInstanceId = _instanceContext.Id
-            };
-
-            _learningStandardsJob.EnqueueJob(jobContext);
-        }
-
-        [HttpPost]
-        public async Task<ActionResult> ResetLearningStandards()
-        {
-            await _resetLearningStandards.Execute();
-
-            return RedirectToActionJson<OdsInstanceSettingsController>(
-                x => x.LearningStandards(),
-                "Credentials successfully reset");
-        }
-
         private async Task<OdsApiConnectionInformation> GetConnectionInformationProvider()
         {
             return await _apiConnectionInformationProvider.GetConnectionInformationForEnvironment();
